Move example page. Remove TagsInput local state. Update typing. Rename vars
