+++ client/src/containers/CreateIssueModal/logic/context.tsx
@@ -39,7 +39,9 @@ export const useCreateIssueModalContext = () => {
 		throw new Error('useCreateIssueModalContext must be used inside a ContextProvider');
 	}
 
-	const set = (key: string, value: any) =>
+	type Value = string | number | boolean | undefined | string[];
+
+	const set = (key: keyof WebApi.Issue.PartialIssue, value: Value) =>
 		dispatch({
 			type: CHANGE,
 			key,
